refactor: make transport and wire format pluggable
Different transport mechanism can be plugged in by providing an implementation of `AbstractTransport`.
E.g. one for RabbitMQ, or even one using in-built Julia messaging protocol.
An implementation of ZMQTransport is included in the package.

Different message interchange format can be plugged in by providing an implementaton of `AbstractMsgFormat`.
E.g. Julia native serialization, ProtoBuf.
An implementation of JSONMsgFormat is included in the package.