add dict msgformat and inproc transport
`DictMsgFormat` and `InProcTransport` enable simpler and faster
communication when a light workload needs to be hosted by a single process.

There is no RPC or serialization involved in the internal communication.
The overhead of converting data from and to the external request format
(likely HTTP) remains though, and is unavoidable.