allow port reuse on HTTP end
This will allow multiple HTTP servers to be multiplexed on the same port on the HTTP end.
This is useful to bypass the limitation in Julia where a `ccall` blocks all tasks in the process.
With this change multiple HTTP clients can get requests serviced simultaneously using more than one servers.
The HTTP ends can further talk to one or more backends as needed.